Improve StreamlineSignInDefaultScreen
