+++ auth-ui/api/current.api
@@ -60,6 +60,14 @@ package com.google.android.horologist.auth.ui.common.screens.streamline {
   @com.google.android.horologist.auth.ui.ExperimentalHorologistAuthUiApi public abstract sealed class StreamlineSignInDefaultScreenState {
   }
 
+  public static final class StreamlineSignInDefaultScreenState.Idle extends com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState {
+    field public static final com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState.Idle INSTANCE;
+  }
+
+  public static final class StreamlineSignInDefaultScreenState.Loading extends com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState {
+    field public static final com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState.Loading INSTANCE;
+  }
+
   public static final class StreamlineSignInDefaultScreenState.MultipleAccountsAvailable extends com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState {
     ctor public StreamlineSignInDefaultScreenState.MultipleAccountsAvailable(java.util.List<com.google.android.horologist.auth.composables.model.AccountUiModel> accounts);
     method public java.util.List<com.google.android.horologist.auth.composables.model.AccountUiModel> component1();
@@ -70,14 +78,6 @@ package com.google.android.horologist.auth.ui.common.screens.streamline {
 
   public static final class StreamlineSignInDefaultScreenState.NoAccountsAvailable extends com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState {
     field public static final com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState.NoAccountsAvailable INSTANCE;
-  }
-
-  public static final class StreamlineSignInDefaultScreenState.ParentState extends com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState {
-    ctor public StreamlineSignInDefaultScreenState.ParentState(com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInScreenState state);
-    method public com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInScreenState component1();
-    method public com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState.ParentState copy(com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInScreenState state);
-    method public com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInScreenState getState();
-    property public final com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInScreenState state;
   }
 
   public static final class StreamlineSignInDefaultScreenState.SignedIn extends com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState {
@@ -93,9 +93,6 @@ package com.google.android.horologist.auth.ui.common.screens.streamline {
     method public kotlinx.coroutines.flow.StateFlow<com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState> getUiState();
     method public void onAccountSelected(com.google.android.horologist.auth.composables.model.AccountUiModel account);
     method public void onIdleStateObserved();
-    method public void onMultipleAccountsAvailable(java.util.List<com.google.android.horologist.auth.composables.model.AccountUiModel> accounts);
-    method public void onNoAccountsAvailable();
-    method public void onSingleAccountAvailable(com.google.android.horologist.auth.composables.model.AccountUiModel account);
     property public final kotlinx.coroutines.flow.StateFlow<com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInDefaultScreenState> uiState;
   }
